Hacking on kgraph for python3 and modern gdbs.
Pretty unhappy.  Using plot.show() works fine with gdb and a local
target, but with openocd and a remote target, after closing the plot,
and 'c'ontinuing, ctrl-c is no longer handled.

Using plot.show(block=False) makes c/ctrl-c work again, but doesn't
actually show the plot, as nothing starts the event loop.

Saving a file and xdg-opening it is.... gross, but functional for the
time being.